bigdata-service/requirements.txt
aniso8601==8.1.1
api==0.0.7
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
elasticsearch==7.10.1
Flask==1.1.2
Flask-RESTful==0.3.8
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
JsonForm==0.0.2
jsonschema==3.2.0
JsonSir==0.0.2
logger==1.4
MarkupSafe==1.1.1
marshmallow==3.12.1
nose==1.3.7
numpy==1.20.3
packaging==20.9
pandas==1.2.4
pluggy==0.13.1
py==1.10.0
pymongo==3.11.2
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.2
python-dateutil==2.8.1
Python-EasyConfig==0.1.7
pytz==2020.5
PyYAML==5.4.1
requests==2.25.1
Resource==0.2.1
scikit-learn==0.24.2
scipy==1.6.3
six==1.15.0
threadpoolctl==2.1.0
toml==0.10.2
urllib3==1.26.3
Werkzeug==1.0.1
kafka-python>=2.0.2
-e git+https://github.com/rkasralikar/mentorlink/common-modules.git#egg=common_modules
